Converted gallery to a class
