+++ src/php/skaut-google-drive-gallery.php
@@ -72,13 +72,13 @@ require_once __DIR__ . '/exceptions/class-unsupported-value-exception.php';
 require_once __DIR__ . '/frontend/interface-pagination-helper-interface.php';
 require_once __DIR__ . '/frontend/class-api-fields.php';
 require_once __DIR__ . '/frontend/class-block.php';
+require_once __DIR__ . '/frontend/class-gallery.php';
 require_once __DIR__ . '/frontend/class-infinite-pagination-helper.php';
 require_once __DIR__ . '/frontend/class-pagination-helper.php';
 require_once __DIR__ . '/frontend/class-options-proxy.php';
 require_once __DIR__ . '/frontend/class-single-page-pagination-helper.php';
 require_once __DIR__ . '/frontend/shortcode.php';
 require_once __DIR__ . '/frontend/page.php';
-require_once __DIR__ . '/frontend/gallery.php';
 
 require_once __DIR__ . '/admin/google-api-lib.php';
 require_once __DIR__ . '/admin/class-settings-pages.php';
@@ -96,7 +96,7 @@ function init() {
 	\Sgdg\Frontend\Shortcode\register();
 	new \Sgdg\Frontend\Block();
 	\Sgdg\Frontend\Page\register();
-	\Sgdg\Frontend\Gallery\register();
+	new \Sgdg\Frontend\Gallery();
 	new \Sgdg\Admin\Settings_Pages();
 	new \Sgdg\Admin\TinyMCE_Plugin();
 }
